add currentChain in WalletServices
